requirements.txt
beautifulsoup4==4.12.2
discord.py==1.6.0
django-extensions==3.2.3
django-ratelimit==4.1.0
Markdown==3.5.2
pillow==10.2.0
pip==23.3.2
psycopg2==2.9.9
python-memcached==1.61
setuptools==69.0.3
uwuify==1.3.0
wheel==0.42.0